[IMP] hr : revamp bank form
-  add bank details ona page in the account form
- add employee field that refreance the partner employee if found

Task: 4531266
