+++ addons/hr/__manifest__.py
@@ -38,6 +38,7 @@
         'views/hr_views.xml',
         'views/res_config_settings_views.xml',
         'views/res_partner_views.xml',
+        'views/res_partner_bank_views.xml',
         'views/discuss_channel_views.xml',
         'views/res_users.xml',
         'views/hr_templates.xml',
