+++ addons/hr/models/__init__.py
@@ -19,5 +19,6 @@ from . import res_config_settings
 from . import res_users
 from . import res_company
 from . import res_partner
+from . import res_partner_bank
 from . import resource
 from . import ir_ui_menu
