+++ addons/hr/models/res_partner.py
@@ -62,16 +62,3 @@ class ResPartner(models.Model):
         employees = {employee for [employee] in employee_data}
         for partner in self:
             partner.employee = partner in employees
-
-
-class ResPartnerBank(models.Model):
-    _inherit = 'res.partner.bank'
-
-    def _compute_display_name(self):
-        account_employee = self.browse()
-        if not self.env.user.has_group('hr.group_hr_user'):
-            account_employee = self.sudo().filtered("partner_id.employee_ids")
-            for account in account_employee:
-                account.sudo(self.env.su).display_name = \
-                    account.acc_number[:2] + "*" * len(account.acc_number[2:-4]) + account.acc_number[-4:]
-        super(ResPartnerBank, self - account_employee)._compute_display_name()
